fix: driver crd daemonset volume setup
This commit fixes an issue that is only present in the Driver CRD
created Daemonset. Basically, the daemonset spec is being extended with
the additionalConfigs volumes - these volumes contain references to the
repoconfig, certconfig etc - these volumes are ConfigMaps. However, in
some cases, there are by default additional volumes that are not
ConfigMaps but HostPaths, ie. in the case of rhel driver images the
subsriptiob-config-%d ones. These cause a failure during generation of
the daemonset spec as the templating engine expects a ConfigMap object.